refactor(sweep): move the quiet-round rule beside its window, out of job-state
job-state.ts and its test may shrink but never grow; the quiet rule and its
tests pushed both past their ceiling. The rule belongs with the window it
reads anyway, so platform/quiet-round.ts now holds quietRoundDays,
lastRoundActivityAt and shouldAutoAbandonQuietRound together, with their
own tests. job-state returns to its baseline.

self-build.test.ts keeps the end-to-end sweep case and is bumped: its app
factory and auth helpers are local to that file.
