+++ eslint-rules/comment-prose-baseline.json
@@ -137,8 +137,8 @@
     "apps/api/src/creation/job-admin-routes.ts": 792,
     "apps/api/src/creation/job-costs.test.ts": 230,
     "apps/api/src/creation/job-costs.ts": 805,
-    "apps/api/src/creation/job-state.test.ts": 506,
-    "apps/api/src/creation/job-state.ts": 2132,
+    "apps/api/src/creation/job-state.test.ts": 480,
+    "apps/api/src/creation/job-state.ts": 2067,
     "apps/api/src/creation/module-size.ts": 128,
     "apps/api/src/creation/owner-games.ts": 98,
     "apps/api/src/creation/quota-gate.ts": 223,
@@ -268,6 +268,8 @@
     "apps/api/src/platform/openai-apps-challenge.test.ts": 62,
     "apps/api/src/platform/openai-apps-challenge.ts": 181,
     "apps/api/src/platform/quiet-round-days.ts": 15,
+    "apps/api/src/platform/quiet-round.test.ts": 44,
+    "apps/api/src/platform/quiet-round.ts": 41,
     "apps/api/src/platform/quota-peek.ts": 223,
     "apps/api/src/platform/rate-limit.ts": 59,
     "apps/api/src/platform/seal-preview.ts": 58,
